chore: address PR comments
